feature-policy/experimental-features/layout-animations-disabled-violation-report-css-tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/feature-policy/experimental-features/resources/common.js"></script>
<title> 'layout-animations' Policy : violation reports from CSS
</title>
<script>
  // Sanity-check: Feature policy 'layout-animations' is disabled.
  test(() => {
    document.policy.allowedFeatures().forEach((enabled_feature) => {
      assert_not_equals(enabled_feature, "layout-animations");
    });
  },
  "Verify 'layout-animations' is not in document's feature list.");

  // Verify reports are generated for linked stylesheets.
  promise_test(async () => {
    let promise = wait_for_violation_in_file(
      "layout-animations", "animation-property-left.css");
    let link = document.createElement("link");
    link.rel = "stylesheet";
    link.href ="/feature-policy/experimental-features/resources/" +
               "animation-property-left.css";
    document.querySelector("head").appendChild(link);
    await promise;
  },
  "Verify linked stylesheet with banned animation generates violation " +
  "reports.");

  // Verify reports are generated for inline stylesheets.
  promise_test(async () => {
    let promise = wait_for_violation_in_file(
      "layout-animations",
      "layout-animations-disabled-violation-report-css-tentative.html");
    let style = document.createElement("style");
    style.innerHTML = `@keyframes animation_property_top {
                          from { top: 0px }
                          to { top: 100px }
                      }`;
    document.body.appendChild(style);
    await promise;
  },
  "Verify inline style with banned animation generates violation reports.");
</script>
</body>
